% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{ensemblQueryGetPops}
\alias{ensemblQueryGetPops}
\title{Function to get list of populations that Ensembl has available to query LD.}
\usage{
ensemblQueryGetPops()
}
\value{
data.frame of populations.
}
\description{
Function to get list of populations that Ensembl has available to query LD.
}
\examples{
ensemblQueryGetPops()

}
