% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblQueryLD.R
\name{ensemblQueryLDwithSNP}
\alias{ensemblQueryLDwithSNP}
\title{Function to query Ensembl LD data with a single rsID}
\usage{
ensemblQueryLDwithSNP(
  rsid,
  r2 = 0.8,
  d.prime = 0.8,
  window.size = 500,
  pop = "1000GENOMES:phase_3:EUR"
)
}
\arguments{
\item{pop}{}
}
\value{
data.frame with 5 columns:
`query` (original query SNP),
`snp_in_ld` (all SNPs in LD within parameters specified),
`r2` (the r-squared value for query-snp_in_ld),
`d_prime` (the D' value for query-snp_in_ld),
`population_name` (the population LD was calculated from)
}
\description{
Function to query Ensembl LD data with a single rsID
}
\examples{
query.ensemble.ld.with.snp(rsid="rs3851179", r2=0.8, d.prime=0.8, window.size=500, pop="1000GENOMES:phase_3:EUR")
}
