% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblQueryLD.R
\name{ensemblQueryLDwithSNPlist}
\alias{ensemblQueryLDwithSNPlist}
\title{Function to apply `ensemblQueryLDwithSNP` to a list of SNPs of length<1000
Outputs a table containing all SNPs in LD with all query SNPs in rsid.list.
Outputs NA row if no SNPs in LD found.}
\usage{
ensemblQueryLDwithSNPlist(
  rsid.list,
  r2 = 0.8,
  d.prime = 0.8,
  window.size = 500,
  pop = "1000GENOMES:phase_3:EUR"
)
}
\arguments{
\item{pop}{}
}
\description{
Function to apply `ensemblQueryLDwithSNP` to a list of SNPs of length<1000
Outputs a table containing all SNPs in LD with all query SNPs in rsid.list.
Outputs NA row if no SNPs in LD found.
}
\examples{

ensemblQueryLDwithSNPlist(rsid.list=c("rs7153434","rs1963154","rs12672022","rs3852802","rs12324408","rs56346870"),
                          r2=0.8,
                          d.prime=0.8,
                          window.size=500,
                          pop="1000GENOMES:phase_3:EUR")

}
