% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblQueryLDpairEndpoint.R
\name{ensemblQueryLDwithSNPpair}
\alias{ensemblQueryLDwithSNPpair}
\title{Function to query Ensembl LD data with a pair of rsIDs.
This function will return r-squared and D' values for the rsID pair.}
\usage{
ensemblQueryLDwithSNPpair(rsid1, rsid2, pop = "1000GENOMES:phase_3:EUR")
}
\arguments{
\item{rsid1}{String. Variant ID 1.}

\item{rsid2}{String. Variant ID 2.}

\item{pop}{String. Population for which to compute LD. Use `ensemblQueryGetPops()` to retrieve a list of all populations with LD data. Default is 1000GENOMES:phase_3:EUR.}
}
\value{
data.frame with 5 columns.
}
\description{
Function to query Ensembl LD data with a pair of rsIDs.
This function will return r-squared and D' values for the rsID pair.
}
\examples{
ensemblQueryLDwithSNPpair(
  rsid1="rs6792369",
  rsid2="rs1042779",
  pop="1000GENOMES:phase_3:EUR"
)

}
