% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblQueryLDwindowEndpoint.R
\name{ensemblQueryLDwithSNPwindowList}
\alias{ensemblQueryLDwithSNPwindowList}
\title{Function to apply `ensemblQueryLDwithSNPwindow` to a list of SNPs of length<1000
Outputs a table containing all SNPs in LD with all query SNPs in rsid.list.
Outputs NA row if no SNPs in LD found.}
\usage{
ensemblQueryLDwithSNPwindowList(
  rsid.list,
  r2 = 0.8,
  d.prime = 0.8,
  window.size = 500,
  pop = "1000GENOMES:phase_3:EUR"
)
}
\arguments{
\item{rsid.list}{Character vector of variant IDs.}

\item{r2}{Float. Measure of LD. If r-squared is provided only return pairs of variants whose r-squared value is equal to or greater than the value provided.}

\item{d.prime}{Float. Measure of LD. If D' is provided only return pairs of variants whose D' value is equal to or greater than the value provided.}

\item{window.size}{Integer. Window size in kb. The maximum allowed value for the window size is 500 kb. LD is computed for the given variant and all variants that are located within the specified window.}

\item{pop}{String. Population for which to compute LD. Use `ensemblQueryGetPops()` to retrieve a list of all populations with LD data. Default is 1000GENOMES:phase_3:EUR.}
}
\description{
Function to apply `ensemblQueryLDwithSNPwindow` to a list of SNPs of length<1000
Outputs a table containing all SNPs in LD with all query SNPs in rsid.list.
Outputs NA row if no SNPs in LD found.
}
\examples{
ensemblQueryLDwithSNPwindowList(rsid.list=c("rs7153434","rs1963154","rs12672022","rs3852802","rs12324408","rs56346870"),
                          r2=0.8,
                          d.prime=0.8,
                          window.size=500,
                          pop="1000GENOMES:phase_3:EUR")

}
