% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblQueryR-package.R
\docType{package}
\name{ensemblQueryR-package}
\alias{ensemblQueryR}
\alias{ensemblQueryR-package}
\title{ensemblQueryR: Simple, Fast and Efficient Querying of the Ensembl API}
\description{
Simple, Fast and Efficient Querying of the Ensembl API.
}
\keyword{internal}
