% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{pingEnsembl}
\alias{pingEnsembl}
\title{Function to check whether the Ensembl server is up by pinging it.}
\usage{
pingEnsembl()
}
\value{
Integer. Where 1 is indicative of a successful ping.
}
\description{
Function to check whether the Ensembl server is up by pinging it.
}
\examples{
pingEnsembl()

}
