% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblQueryR-package.R
\docType{package}
\name{ensemblQueryR-package}
\alias{ensemblQueryR}
\alias{ensemblQueryR-package}
\title{ensemblQueryR: Simple, Fast and Efficient Querying of the 'Ensembl' API}
\description{
Suite of tools to enable fast, efficient, high-throughput, R workflow integrable querying of 'Ensembl' REST API endpoints (\url{https://rest.ensembl.org/}).
}
\keyword{internal}
